added ability to autocomplete city when searching for a city to visit
