+++ app/controllers/trips_controller.rb
@@ -26,7 +26,7 @@ class TripsController < ApplicationController
   def create
     @trip = Trip.new(trip_params)
     @trip.user = current_user
-    @trip.location = (@trip.location.split(' ').map { |e| e.capitalize! }).join(' ')
+    @trip.location = @trip.location.split(",")[0]
     @trip.status = "saved"
     if @trip.save
       redirect_to trip_path(@trip)
