+++ app/javascript/packs/application.js
@@ -1,17 +1,10 @@
-import { Application } from "stimulus"
-import { definitionsFromContext } from "stimulus/webpack-helpers"
-
-const application = Application.start()
-const context = require.context("controllers", true, /.js$/)
-application.load(definitionsFromContext(context))
-
 import "bootstrap";
 
 import "controllers"
 
 window.initMap = function (...args) {
   const event = document.createEvent("Events")
-  event.initEvent("google-maps-callback", true, true)
+  event.initEvent("google-search-callback", true, true)
   event.args = args
   window.dispatchEvent(event)
 }
